- Correct fix for table ezurlalias_ml, started in rev. 6
git-svn-id: http://svn.projects.ez.no/ezoracle/trunk@9 436571c7-bdc5-44d1-b1ef-e55a79f94445

+++ extension/ezoracle/update/database/ezpublish/4.1/unstable/dbupdate-4.0.0-to-4.1.0alpha1.sql
@@ -17,4 +17,4 @@ CREATE UNIQUE INDEX ezcontentobject_remote_id ON ezcontentobject (remote_id);
 CREATE UNIQUE INDEX ezgeneral_digest_user_sett_add on ezgeneral_digest_user_settings(address);
 DELETE FROM ezgeneral_digest_user_settings WHERE address not in (SELECT email FROM ezuser);
 
-ALTER TABLE ezurlalias_ml ADD alias_redirects int(11) NOT NULL default 1;
+ALTER TABLE ezurlalias_ml ADD alias_redirects number(11) NOT NULL default 1;
